update to current progress
Former-commit-id: bed7f1bddb7af09c0f288c1150c6b1c5c106e3be [formerly 02402a0b62d22bfba17d6dfef2f443c1500b450a] [formerly 93bff37d83bad8297647c88c0b42afe81ed277e7 [formerly 08274c78a878dba3fa13551e009ca7aba6528beb]]
Former-commit-id: f66285070e26f0193e87c728c4a0d3913e9a6868 [formerly 85e3eb7005d566d372945751835e29205ff816b4]
Former-commit-id: deff595044d44ac83b2a5c920a64abb2d97054f5

+++ tools/1172compress.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
-echo -n -e \\x11\\x72 > header.tmp
-cat $1| gzip --no-name --best | tail --bytes=+11 | head --bytes=-8 > $1.tmp
+if [ ! -f header.tmp ]; then
+    echo -n -e \\x11\\x72 > header.tmp
+fi
+cat $1| tools/gzip --no-name --best | tail --bytes=+11 | head --bytes=-8 > $1.tmp
 cat header.tmp $1.tmp > $2
-rm header.tmp $1.tmp
+rm $1.tmp
